pyproject.toml
[tool.poetry]
name = "chatzeus"
version = "0.1.0"
description = ""
authors = ["django-iori <iori4321999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-cors-headers = "^4.2.0"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
boto3 = "^1.28.59"
watchtower = "^3.0.1"
djangorestframework = "^3.14.0"
markdown = "^3.5"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
